refactor: re-structure core functionality placement
