+++ backend/src/core/index.ts
@@ -1,3 +1,2 @@
-export * from './objects/files/index.js'
 export * from './objects/index.js'
 export * from './users/index.js'
