+++ backend/src/core/objects/index.ts
@@ -2,4 +2,5 @@ export * from './interactions.js'
 export * from './object.js'
 export * from './ownership.js'
 export * from './uploadStatus.js'
+export * from './files/index.js'
 export * from './nodes.js'
