+++ packages/models/src/uploads/upload.ts
@@ -1,23 +1,23 @@
-import { z } from 'zod'
-import { FolderTreeFolder, FolderTreeFolderSchema } from '../objects/index.js'
+import { z } from "zod";
+import { FolderTreeFolder, FolderTreeFolderSchema } from "../objects/index.js";
 import {
   CompressionAlgorithm,
   EncryptionAlgorithm,
   FileUploadOptions,
   OffchainFileMetadata,
   OffchainFolderMetadata,
-} from '@autonomys/auto-dag-data'
+} from "@autonomys/auto-dag-data";
 
 export enum UploadType {
-  FILE = 'file',
-  FOLDER = 'folder',
+  FILE = "file",
+  FOLDER = "folder",
 }
 
 export enum UploadStatus {
-  PENDING = 'pending',
-  MIGRATING = 'migrating',
-  CANCELLED = 'cancelled',
-  FAILED = 'failed',
+  PENDING = "pending",
+  MIGRATING = "migrating",
+  CANCELLED = "cancelled",
+  FAILED = "failed",
 }
 
 export type UploadEntry = {
@@ -47,11 +47,11 @@ export const uploadOptionsSchema = z.object({
       chunkSize: z.number().optional(),
     })
     .optional(),
-})
+});
 
 // & operator is used for ensure it represents a FileUploadOptions object
 export type UploadOptions = z.infer<typeof uploadOptionsSchema> &
-  FileUploadOptions
+  FileUploadOptions;
 
 export const fileUploadSchema = z.object({
   id: z.string(),
@@ -64,11 +64,11 @@ export const fileUploadSchema = z.object({
   mimeType: z.string().nullable(),
   oauthProvider: z.string(),
   oauthUserId: z.string(),
-})
+});
 
 export type FileUpload = z.infer<typeof fileUploadSchema> & {
-  uploadOptions: FileUploadOptions | null
-}
+  uploadOptions: FileUploadOptions | null;
+};
 
 export const mapModelToTable = (upload: Upload): UploadEntry => {
   return {
@@ -83,8 +83,8 @@ export const mapModelToTable = (upload: Upload): UploadEntry => {
     oauth_provider: upload.oauthProvider,
     oauth_user_id: upload.oauthUserId,
     upload_options: upload.uploadOptions,
-  }
-}
+  };
+};
 
 export const folderUploadSchema = z.object({
   id: z.string(),
@@ -98,18 +98,18 @@ export const folderUploadSchema = z.object({
   oauthProvider: z.string(),
   oauthUserId: z.string(),
   uploadOptions: z.null(),
-})
+});
 
-export type FolderUpload = z.infer<typeof folderUploadSchema>
-export type Upload = FileUpload | FolderUpload
+export type FolderUpload = z.infer<typeof folderUploadSchema>;
+export type Upload = FileUpload | FolderUpload;
 
-export type UploadArtifacts = FileArtifacts | FolderArtifacts
+export type UploadArtifacts = FileArtifacts | FolderArtifacts;
 
 export type FileArtifacts = {
-  metadata: OffchainFileMetadata
-}
+  metadata: OffchainFileMetadata;
+};
 
 export type FolderArtifacts = {
-  metadata: OffchainFolderMetadata
-  childrenArtifacts: UploadArtifacts[]
-}
+  metadata: OffchainFolderMetadata;
+  childrenArtifacts: UploadArtifacts[];
+};
